pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abfahrt"
version = "0.1.0"
description = "BVG Train Departure Display"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pygame>=2.5.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
hardware = [
    "luma.oled>=3.12.0",
]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
